dlinknet uses mindcv resnet backbone
