+++ official/cv/dlinknet/models/dlinknet.py
@@ -1,11 +1,74 @@
+import math
+import numpy as np
 from functools import partial
+import mindcv
 
 from mindspore import nn, load_checkpoint, load_param_into_net
 from mindspore.common.tensor import Tensor
 from mindspore.mint.nn.functional import relu, batch_norm
 from mindspore.mint import sigmoid
-from models.resnet import resnet34, resnet50, kaiming_uniform
 from config import parse_args
+
+
+def calculate_gain(nonlinearity, param=None):
+    """calculate_gain"""
+    linear_fns = ['linear', 'conv1d', 'conv2d', 'conv3d', 'conv_transpose1d', 'conv_transpose2d', 'conv_transpose3d']
+    res = 0
+    if nonlinearity in linear_fns or nonlinearity == 'sigmoid':
+        res = 1
+    elif nonlinearity == 'tanh':
+        res = 5.0 / 3
+    elif nonlinearity == 'relu':
+        res = math.sqrt(2.0)
+    elif nonlinearity == 'leaky_relu':
+        if param is None:
+            negative_slope = 0.01
+        elif not isinstance(param, bool) and isinstance(param, int) or isinstance(param, float):
+            # True/False are instances of int, hence check above
+            negative_slope = param
+        else:
+            raise ValueError("negative_slope {} not a valid number".format(param))
+        res = math.sqrt(2.0 / (1 + negative_slope ** 2))
+    else:
+        raise ValueError("Unsupported nonlinearity {}".format(nonlinearity))
+    return res
+
+
+def _calculate_fan_in_and_fan_out(tensor):
+    """_calculate_fan_in_and_fan_out"""
+    dimensions = len(tensor)
+    if dimensions < 2:
+        raise ValueError("Fan in and fan out can not be computed for tensor with fewer than 2 dimensions")
+    if dimensions == 2:  # Linear
+        fan_in = tensor[1]
+        fan_out = tensor[0]
+    else:
+        num_input_fmaps = tensor[1]
+        num_output_fmaps = tensor[0]
+        receptive_field_size = 1
+        if dimensions > 2:
+            receptive_field_size = tensor[2] * tensor[3]
+        fan_in = num_input_fmaps * receptive_field_size
+        fan_out = num_output_fmaps * receptive_field_size
+    return fan_in, fan_out
+
+
+def _calculate_correct_fan(tensor, mode):
+    mode = mode.lower()
+    valid_modes = ['fan_in', 'fan_out']
+    if mode not in valid_modes:
+        raise ValueError("Mode {} not supported, please use one of {}".format(mode, valid_modes))
+    fan_in, fan_out = _calculate_fan_in_and_fan_out(tensor)
+    return fan_in if mode == 'fan_in' else fan_out
+
+
+def kaiming_uniform(inputs_shape, a=0., mode='fan_in', nonlinearity='leaky_relu'):
+    fan = _calculate_correct_fan(inputs_shape, mode)
+    gain = calculate_gain(nonlinearity, a)
+    std = gain / math.sqrt(fan)
+    bound = math.sqrt(3.0) * std  # Calculate uniform bounds from standard deviation
+    return np.random.uniform(-bound, bound, size=inputs_shape).astype(np.float32)
+
 
 conv_weight_init = partial(kaiming_uniform, mode="fan_out", nonlinearity='relu')
 
@@ -98,15 +161,14 @@ class DLinkNet34(nn.Cell):
         self.args = parse_args()
 
         filters = [64, 128, 256, 512]
-        resnet = resnet34(1001)
         if use_backbone:
-            param_dict = load_checkpoint(self.args.pretrained_ckpt)
-            load_param_into_net(resnet, param_dict)
+            resnet = mindcv.create_model('resnet34', pretrained=False, checkpoint_path=self.args.pretrained_ckpt)
+        else:
+            resnet = mindcv.create_model('resnet34', pretrained=False, checkpoint_path='')
 
         self.firstconv = resnet.conv1
         self.firstbn = resnet.bn1
-        self.pad = resnet.pad
-        self.firstmaxpool = resnet.maxpool
+        self.firstmaxpool = resnet.max_pool
         self.encoder1 = resnet.layer1
         self.encoder2 = resnet.layer2
         self.encoder3 = resnet.layer3
@@ -133,7 +195,6 @@ class DLinkNet34(nn.Cell):
         x = self.firstconv(x)
         x = self.firstbn(x)
         x = relu(x)
-        x = self.pad(x)
         x = self.firstmaxpool(x)
         e1 = self.encoder1(x)
         e2 = self.encoder2(e1)
@@ -164,14 +225,14 @@ class DLinkNet50(nn.Cell):
         self.args = parse_args()
 
         filters = [256, 512, 1024, 2048]
-        resnet = resnet50(1001)
         if use_backbone:
-            param_dict = load_checkpoint(self.args.pretrained_ckpt)
-            load_param_into_net(resnet, param_dict)
+            resnet = mindcv.create_model('resnet50', pretrained=False, checkpoint_path=self.args.pretrained_ckpt)
+        else:
+            resnet = mindcv.create_model('resnet50', pretrained=False, checkpoint_path='')
 
         self.firstconv = resnet.conv1
         self.firstbn = resnet.bn1
-        self.firstmaxpool = resnet.maxpool
+        self.firstmaxpool = resnet.max_pool
         self.encoder1 = resnet.layer1
         self.encoder2 = resnet.layer2
         self.encoder3 = resnet.layer3
