+++ official/cv/dlinknet/requirements.txt
@@ -2,3 +2,4 @@ opencv-python >= 4.1.2.30
 numpy
 pyyaml
 scikit-learn
+mindcv == 0.3.0
